restful_sctr/requirement.txt
appdirs==1.4.0
Django==1.10.5
djangorestframework==3.5.3
numpy==1.12.0
packaging==16.8
pandas==0.19.2
pkg-resources==0.0.0
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
six==1.10.0
